pyproject.toml
[project]
name = "slidegenie"
version = "0.1.0"
description = "Local AI slide generation tool based on ppt-addin/backend pipeline"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.55.0",
    "google-auth>=2.43.0",
    "python-pptx>=0.6.23",
    "pillow>=11.0.0",
    "jinja2>=3.1.6",
    "numpy>=2.2.0",
    "click>=8.1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slidegenie"]
